% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/_PACKAGE.R
\docType{package}
\name{unimelbdown-package}
\alias{unimelbdown-package}
\title{unimelbdown: University of Melbourne R Markdown thesis template}
\description{
Provides functions to render PDF, Gitbook, Word, and EPUB theses
for the University of Melbourne using bookdown.
}
\author{
\strong{Maintainer}: Luke Zappia \email{lazappi@users.noreply.github.com}

Authors:
\itemize{
  \item Chester Ismay \email{chester.ismay@gmail.com}
  \item Nick Solomon \email{nick@nicksolomon.me}
}

}
